Added buttons and a ; in each string
Saw that the script was sending multiple messages for each button push (which has been fixed on the raspberry pi), but added a semi-colon at the end of each message string for parsing purposes.  Also, i added MAC addresses for 3 new buttons.  Note that poof, sky, and kitty do not always trigger a message.  i'm not sure if it is because it is a newer model of the dash buttons or not.  but exercise caution when expecting them to work consistently.